Misc fixes to Geometry and Line etc.

+++ android/app/src/main/java/thothbot/parallax/core/shared/materials/LineBasicMaterial.java
@@ -42,7 +42,7 @@ public class LineBasicMaterial extends Material
 	
 	private Material.COLORS vertexColors;
 	
-	private double linewidth;
+	private float linewidth;
 	
 	public LineBasicMaterial()
 	{	
@@ -50,7 +50,7 @@ public class LineBasicMaterial extends Material
 		
 		this.color = new Color(0xffffff);
 		
-		setLinewidth(1.0);
+		setLinewidth(1.0f);
 		
 		setVertexColors(Material.COLORS.NO);
 	}
@@ -65,7 +65,7 @@ public class LineBasicMaterial extends Material
 	 * Line thickness. Default is 1.
 	 * @return
 	 */
-	public double getLinewidth() {
+	public float getLinewidth() {
 		return this.linewidth;
 	}
 	
@@ -75,7 +75,7 @@ public class LineBasicMaterial extends Material
 	 * Due to limitations in the <a href="https://code.google.com/p/angleproject/">ANGLE layer</a>, on Windows platforms linewidth will always be 1 regardless of the set value.
 	 * @param linewidth
 	 */
-	public void setLinewidth(double linewidth) {
+	public void setLinewidth(float linewidth) {
 		this.linewidth = linewidth;
 	}
